--- pkg/sources/ipfs.go
@@ -1,5 +1,0 @@
-package sources
-
-/*
-IPFS Implementation of sources types
-*/
